+++ Mixin/Service/DeviceTransfer/Model/DeviceTransferDisplayState.swift
@@ -6,7 +6,7 @@ enum DeviceTransferDisplayState {
     case ready
     case connected
     case transporting(processedCount: Int, totalCount: Int)
-    case importing(Double)
+    case importing(Float)
     case failed(DeviceTransferConnectionClosedReason)
     case finished
     case closed
